Remove hard-coded module list from commands-index
Expected review fix:
https://github.com/joewalker/gecko-dev/commit/a7a75328845edb28aa6426cb6dd6be71870e244e

This splits main.js so the definitions are in definitions.js.

We can't load gDevTools in a child process so the command modules that
load it need to lazyLoad it instead, which is fine because they never
actually need it in the child process.

I've not yet worked out why we need to require(Services) in
console-output. We require it in enough other places that I don't see
the problem with requiring it, but I still don't get it.

It's used here:
https://dxr.mozilla.org/mozilla-central/source/browser/devtools/webconsole/console-output.js#30
And should be defined here:
https://dxr.mozilla.org/mozilla-central/source/toolkit/devtools/Loader.jsm?from=Loader.jsm#31
